+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.0.11] - 2021-5-1
+
+Update form handling for changes in @hotwired/turbo 7.0.0-beta.2:
+
+  - **TurboStreamFormMixin** class and supporting classes removed
+  - **TurboFormMixin** class added that just returns a 422 response on invalid
+  - **Turbo
+
 ## [0.0.10] - 2020-12-30
 
 Remove __str__ methods from TurboStream and TurboFrame classes